requirements.txt
firebase-admin
flask
pyrebase4
pycryptodome